+++ src/core/UTxMultiOperation.pas
@@ -427,12 +427,9 @@ begin
     If (w>0) then begin
       for i:=0 to w-1 do begin
         txr := CT_TMultiOpReceiver_NUL;
-        txr.AccountEPASA.Clear;
         stream.Read(txr.Account,SizeOf(txr.Account));
         stream.Read(txr.Amount,SizeOf(txr.Amount));
         LoadOperationPayloadFromStream(stream,txr.Payload);
-        //
-        txr.AccountEPASA:=TAccountComp.DecodeEPASAPartial(txr.Account,txr.Payload.payload_raw,txr.Payload.payload_type,TEPasa.Empty);
         //
         txreceivers[i] := txr;
       end;